+++ targets/TARGET_STM/TARGET_STM32F1/common_objects.h
@@ -68,6 +68,21 @@ struct serial_s {
 #endif
 };
 
+struct spi_s {
+    SPI_HandleTypeDef handle;
+    IRQn_Type spiIRQ;
+    SPIName spi;
+    PinName pin_miso;
+    PinName pin_mosi;
+    PinName pin_sclk;
+    PinName pin_ssel;
+#ifdef DEVICE_SPI_ASYNCH
+    uint32_t event;
+    uint8_t module;
+    uint8_t transfer_type;
+#endif
+};
+
 #include "gpio_object.h"
 
 #ifdef __cplusplus
