+++ targets/TARGET_STM/TARGET_STM32F3/TARGET_NUCLEO_F303ZE/objects.h
@@ -77,20 +77,6 @@ struct serial_s {
     PinName pin_rx;
 };
 
-struct spi_s {
-    SPIName spi;
-    uint32_t bits;
-    uint32_t cpol;
-    uint32_t cpha;
-    uint32_t mode;
-    uint32_t nss;
-    uint32_t br_presc;
-    PinName pin_miso;
-    PinName pin_mosi;
-    PinName pin_sclk;
-    PinName pin_ssel;
-};
-
 struct i2c_s {
     I2CName  i2c;
     uint32_t slave;
