+++ targets/TARGET_STM/TARGET_STM32L0/TARGET_NUCLEO_L031K6/objects.h
@@ -60,20 +60,6 @@ struct analogin_s {
     uint32_t channel;
 };
 
-struct spi_s {
-    SPIName spi;
-    uint32_t bits;
-    uint32_t cpol;
-    uint32_t cpha;
-    uint32_t mode;
-    uint32_t nss;
-    uint32_t br_presc;
-    PinName  pin_miso;
-    PinName  pin_mosi;
-    PinName  pin_sclk;
-    PinName  pin_ssel;
-};
-
 struct i2c_s {
     I2CName  i2c;
 };
