+++ targets/TARGET_STM/TARGET_STM32L1/TARGET_NUCLEO_L152RE/objects.h
@@ -65,20 +65,6 @@ struct dac_s {
     PinName pin;
 };
 
-struct spi_s {
-    SPIName spi;
-    uint32_t bits;
-    uint32_t cpol;
-    uint32_t cpha;
-    uint32_t mode;
-    uint32_t nss;
-    uint32_t br_presc;
-    PinName  pin_miso;
-    PinName  pin_mosi;
-    PinName  pin_sclk;
-    PinName  pin_ssel;
-};
-
 struct i2c_s {
     I2CName  i2c;
     uint32_t slave;
